requirements.txt
blinker==1.6.2
click==8.1.5
Flask==2.3.2
Flask-Cors==4.0.0
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
PyJWT==1.7.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
waitress==2.1.2
wakeonlan==3.0.0
Werkzeug==2.3.6
